Fix map and peformance optimizations

+++ src/wolfenstein/port/wl_state.h
@@ -18,7 +18,9 @@ void SelectDodgeDir (AActor *ob);
 void SelectRunDir (AActor *ob);
 void SelectWanderDir (AActor *ob);
 bool MoveObj (AActor *ob, int32_t move);
-bool SightPlayer (AActor *ob, double minseedist, double maxseedist, double maxheardist, double fov, const Frame *state);
+// float (not double) on purpose: rv32imafc has hardware single-precision
+// only; these run per enemy per tic.
+bool SightPlayer (AActor *ob, float minseedist, float maxseedist, float maxheardist, float fov, const Frame *state);
 
 void    DamageActor (AActor *ob, AActor *attacker, unsigned damage);
